[zstd_stream] Return an error in Writer.Close if a write fails
Previously, if there was any remaining data to be written to the
underlying Writer (which was almost always the case, since zstd tries to
buffer blocks and frames as much as possible), and there was an
underlying error when doing a Write on the underlying io.Writer, no
error was returned from Close.

This was an issue because clients would believe the stream was correctly
written in its entirety even though the last bytes were not written.

This fixes the issue by catching these errors, closing the zstd stream
and returning the i/o error.