[zstd_stream] Add Writer.Flush()
There are two levels of buffering in the library:
- zstd buffers its data by not closing zstd blocks until a specific call
  to ZSTD_compressStream2 with ZSTD_e_flush or ZSTD_e_end is made
- the Go wrapper buffers the input data to its own buffer

Previously there was no way to flush these two buffers without closing
the stream.

This adds a way to flush these streams in order for any reader that
wants to read was written to the stream until that point, to do so
without waiting for the entire stream to be closed.