Keep decompressing as long as we have compressed data
ZSTD_decompressStream may consume only part of the src buffer, for
example if the input is a concatenation of two compressed streams. In
this case the second stream will remain undecoded in the input buffer
after the first call, and we need to call decompress routine again.

This makes NewReader behavior closer to Decompress, that handles such
concatenated streams transparently.